refactor: split app.go into per-screen files
- Extract message types to messages.go
- Extract styles, status bar, fitToHeight, viewLoading, viewError to styles.go
- Extract EC2/SSM screen logic to screen_ec2.go
- Extract VPC/Subnet screen logic to screen_vpc.go
- Extract RDS screen logic to screen_rds.go
- Extract Route53 screen logic to screen_route53.go
- Extract context picker logic to screen_context.go
- Reduce app.go from 2,127 to 563 lines (core Model, Update, View dispatch)

Pure file split — no logic or interface changes.
